+++ CHANGELOG.md
@@ -5,8 +5,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- VIMEET-2090 - Improve video conference meeting UX
 - VIMEET-2096 - Add image header to webinar
+
+## [2.26.0] - 2020-05-05
+### Added
+- VIMEET-2090 - Improve video conference meeting UX
+- VIMEET-2078 - Keywords search from Sheet Query Builder
+- VIMEET-2067 - Visio settings
+### Fixed
+- MV-220 - Fix logo error when ordering logo in the purchase tunnel
+- MV-232 - Participant can access to all own sheet video meeting
+- MV-234 - Fix sheet / participant export with dates filters
+- Hotfix - Avoid pending job to be rescheduled
 
 ## [2.25.0] - 2020-03-30
 ### Added